requirements.txt
pyyaml
colour
Pillow
black